fixed navbar buttons for login & logout
