+++ client/src/components/Navigation/Navigation.js
@@ -1,11 +1,11 @@
-import React from "react";
+import React, { useContext, useState }  from "react";
 import { NavLink} from "react-router-dom";
 import { Navbar, Nav, Form, Jumbotron, Container, Row, Col} from "react-bootstrap";
 import Time from '../CurrTime/CurrTimePage'
 import LogOptions from "../LogOptions/LogOptions"
 import bgimage from "./header.png"
-
-
+import UserContext from '../../Context/UserContext';
+import { useHistory } from 'react-router-dom';
 
 const jumbotronStyles = 
 { 
@@ -20,6 +20,24 @@ const jumbotronStyles =
 }
             
 const Navigation = () => {
+  const { userData, setUserData } = useContext(UserContext);
+  const history = useHistory();
+  const register = () => history.push('/register')
+  const login = (e) => {
+    e.preventDefault();
+    history.push('/login')
+  }
+
+  const logout = (e) => {
+    e.preventDefault();
+      setUserData({
+        token: undefined,
+        user: undefined
+      });
+      localStorage.setItem("auth-token", "");
+      history.push('/login')
+    };
+    console.log('USER DATA:', userData)
   return (
     <>
     
@@ -31,9 +49,18 @@ const Navigation = () => {
   <Navbar.Toggle aria-controls="basic-navbar-nav" />
   <Navbar.Collapse id="basic-navbar-nav">
     <Nav className="mx-auto ">
+
     <Nav.Link href="/home"><h4>HOME</h4></Nav.Link>
-      <Nav.Link href="/register"><h4>REGISTER</h4></Nav.Link>
-      <Nav.Link href="/login"><h4>LOGIN</h4></Nav.Link>
+    {userData.user ?(
+      <Nav.Link onClick={logout} href="/"><h4>LOGOUT</h4></Nav.Link>
+    ) :(
+      <>
+      <Nav.Link onClick={register} href="/register"><h4>REGISTER</h4></Nav.Link>
+      <Nav.Link onClick={login} href="/login"> <h4>LOGIN</h4></Nav.Link>
+      </>
+    )} 
+      {/* <Nav.Link href="/register"><h4>REGISTER</h4></Nav.Link>
+      <Nav.Link href="/login"><h4>LOGIN</h4></Nav.Link> */}
       <Nav.Link href="/about"><h4>ABOUT US</h4></Nav.Link>  
        
     </Nav>
